requirements/common.txt
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==25.1
django-storages==1.14.5
Django==5.1.7
djangorestframework-simplejwt==5.5.0
djangorestframework==3.15.2
google-auth==2.38.0
gunicorn==23.0.0
Pillow==11.1.0
psycopg2-binary==2.9.10
requests==2.32.3
sentry-sdk==2.23.1
